Have the Install and configure commands for selenium server to run on the selenium worker server by using our command.

+++ tools/install-seleinum-server/uninstall-selenium-server.sh
@@ -1,12 +1,17 @@
+#!/bin/bash
 
-sudo /etc/init.d/selenium stop
-sudo rm -rf /usr/lib/selenium
+if [ -d "/usr/lib/selenium" ]; then
+  sudo /etc/init.d/selenium stop
+  sudo rm -rf /usr/lib/selenium
 
-echo "\nDeleted selenium.\n"
+  echo "\nDeleted selenium.\n"
 
-sudo rm -rf /var/log/selenium
+  sudo rm -rf /var/log/selenium
 
-echo "Deleted logs.\n"
+  echo "Deleted logs.\n"
 
-sudo update-rc.d -f selenium remove
-echo "Removed selenium server.\n"
+  sudo update-rc.d -f selenium remove
+  echo "Removed selenium server.\n"
+else
+  echo "You can not uninstall Selenium is NOT installed."
+fi
